README: add versioning process and support level
Going forward we should have a known and strict policy about versioning
and set the expectations accordingly. This PR adds the related
information.

Note that I also changed which supports are officially supported by us.
This means replying to issues and handling bug fixes. We should not
maintain older versions or very new versions unless we also support it
on DigitalOcean Kubernetes. Currently we don't have the time and
manpower to support fully all Kubernetes versions.