Assume HTTPS protocol on default TLS port when TLS option is given. (#203)
Assume HTTPS protocol on default TLS port when TLS option is given.

Previously, TLS options (such as the certificate ID or pass-through)
would be ignored silently if either an HTTP* protocol or TLS port
annotation was missing. With this change, we automatically enable the
HTTPS protocol on services exposing the default TLS port 443.

Effectively, this means that users may now omit the TLS port and
protocol annotations for common use cases.