workaround: cluster->lb traffic w/ TLS/ProxyProto
This PR excludes traffic with Layer7 rules from a kube-proxy optimization.
The kube-proxy optimization intercepts and redirects cluster egress traffic
destined for the LB to the backing services, avoiding the DigitalOcean LB
entirely.  If the DO LB is providing L7 services like TLS termination, Proxy
Protocol, or StickySessions this breaks access to LB-hosted services
from within the cluster. This PR opts out of the optimization by returning
the IP address in the Hostname field rather than IP.  The IP will be reported
to the user in the same way via tools like kubectl, but is interpreted differently
by kube-proxy.