Use provider IDs for setting LB droplet targets
Node objects may contain a provider ID of the kind
`digitalocean://<droplet ID>`. At times when we need to specify the set
of target droplets for load-balancer creates and updates, we now check
for the availablity of provider IDs and use them directly without
relying on the local droplets cache.

This change renders the droplets cache obsolete: it was originally
introduced to better support accounts with a large number of droplets.
For such cases, the provider ID should be injected through the kubelet
directly. Consequently, we drop the cache and resort to listing all
droplets from the API only when nodes do not come with a provider ID,
which should be acceptable for smaller accounts.