e2e: Do not poll winding down LB if it was never created
Our end-to-end test cleans up after itself, which includes any created
services to make sure external LBs are deleted as well. To verify that
the delete has properly completed, we try to send requests over the LB
until it starts to return errors (which means the LB must be gone).

It is okay to skip this step if no load balancer IP was able received.