Overwrite service load-balancer ID on mismatch
This change causes the service load-balancer ID annotation to be
overwritten if it does not match the given load-balancer ID.

It is required to properly handle the case where the load-balancer is
deleted manually: in that case, CCM won't be able to look up the
load-balancer and create a new one. Without updating the ID, however,
subsequent reconciliations will continue to use the old ID, fail to find
one, try to create a new but fail because a load-balancer with the given
name (based on the service UUID) already exists. At this point, the
reconciliation attempt will loop endlessly.

We need to break the cycle by annotating the ID of the newly created
load-balancer.