Return error based on built secure forwarding rules on port 443
It turned out that redirection only works on port 443, which is why we
need to be more specific.

Additionally, we should not base our decision on the ports specified in
the annotations because (a) we always add the default secure 443 port to
the list of secure ports we check for emptiness if pass-through or a
certificate are requested; and (b) it is the actually built forwarding
rule that determines if HTTP-to-HTTPx redirection can apply, not the
intention expressed through the annotations. For instance, a user could
specify port 443 in the HTTPS ports annotation quite explicitly but
still miss to have a service on port 443 defined, which the previous
implementation would not mark as an error.

Hence, we loop over the built forwarding rules and check that there is
at least one secure entry protocol defined on port 443. As a minor
benefit, this allows us to invoke getRedirectHTTPToHTTPS a single time
only.