Refactor how we retrieve a load balancer from DO API
This also makes it possible to lookup a load balancer by a legacy name
('a' + service.UID). The name can be changed in the
`do-load-balancer-name` annotation and these changes help to make load
balancer lookup more robust. We also keep the logic used to search for
load balancer in a single place.