ccm/do: relax UDP port duplication validation
This was originally added in #409 due to an upstream bug (https://github.com/kubernetes/kubernetes/issues/39188). It turns out, however, that we and many others misinterpreted the scope of that bug - it really only applies for updating an existing service with a new port definition. It does not apply for creating a new service.

Our customers are affected by this, and this validation is overly restrictive compared to errors the Load Balancers API actually returns during its validation. As such, the validation we do here can be relaxed.