Update forwarding rule logic for REGIONAL_NETWORK
Traffic will land on the worker node destined for the port of the service and not the node port. If there is an LB with port:80 and targetPort:8000. Then traffic will be configured to flow through the NLB on port: 80, will be forwarded to a DOKS worker node with DST IP=VIP and DST port=80, and then kube-proxy/cillium will take care of routing it to an appropriate pod.