Add new annotation for using certificate name when configuring load balancer TLS (#722)
* Add new annotation for using certificate name when configuring load balancer TLS

This will improve the users experience when using a certificate. A name is easier to identify then a UUID. If they are using lets encrypt, the underlying UUID will change on each renewal which causes headaches for users. Certificate names are also now unique so we can identify a certificate by its name even if the underlying UUID has changed.